+++ Source/RapyutaSimulationPlugins/Private/Core/RRActorCommon.cpp
@@ -66,71 +66,66 @@ bool URRSceneInstance::IsValid(bool bIsLogged) const
 // ===================================================================================================================================
 // [FRRActorSpawnInfo] --
 //
-FRRActorSpawnInfo::FRRActorSpawnInfo(const FString& InUniqueName,
-                                     const FTransform& InTransform,
-                                     const TArray<FString>& InMeshMaterialNameList,
-                                     bool bIsStationary,
-                                     bool bIsPhysicsEnabled,
-                                     bool bIsCollisionEnabled)
+FRRActorSpawnInfo::FRRActorSpawnInfo()
 {
-    (*this)(InUniqueName, InTransform, EMPTY_STR, InMeshMaterialNameList, bIsStationary, bIsPhysicsEnabled, bIsCollisionEnabled);
+    bIsTickEnabled = false;
+    bIsStationary = false;
+    bIsPhysicsEnabled = true;
+    bIsCollisionEnabled = true;
+    bIsSelfCollision = false;
 }
 
-FRRActorSpawnInfo::FRRActorSpawnInfo(const FString& InUniqueName,
-                                     const FTransform& InTransform,
-                                     const FString& InMeshUniqueName,
-                                     const TArray<FString>& InMeshMaterialNameList,
-                                     bool bIsStationary,
-                                     bool bIsPhysicsEnabled,
-                                     bool bIsCollisionEnabled)
-{
-    (*this)(
-        InUniqueName, InTransform, InMeshUniqueName, InMeshMaterialNameList, bIsStationary, bIsPhysicsEnabled, bIsCollisionEnabled);
-}
-
-FRRActorSpawnInfo::FRRActorSpawnInfo(const FString& InUniqueName,
-                                     const FTransform& InTransform,
+FRRActorSpawnInfo::FRRActorSpawnInfo(const FString& InEntityModelName,
+                                     const FString& InUniqueName,
+                                     const FTransform& InActorTransform,
+                                     const TArray<FTransform>& InMeshRelTransformList,
                                      const TArray<FString>& InMeshUniqueNameList,
-                                     const TArray<FString>& InMeshMaterialNameList,
-                                     bool bIsStationary,
-                                     bool bIsPhysicsEnabled,
-                                     bool bIsCollisionEnabled)
-    : UniqueName(InUniqueName),
-      Transform(InTransform),
+                                     const TArray<FString>& InMaterialNameList,
+                                     bool bInIsStationary,
+                                     bool bInIsPhysicsEnabled,
+                                     bool bInIsCollisionEnabled)
+    : EntityModelName(InEntityModelName),
+      UniqueName(InUniqueName),
+      ActorTransform(InActorTransform),
+      MeshRelTransformList(InMeshRelTransformList),
       MeshUniqueNameList(InMeshUniqueNameList),
-      MeshMaterialNameList(InMeshMaterialNameList),
-      IsStationary(bIsStationary),
-      IsPhysicsEnabled(bIsPhysicsEnabled),
-      IsCollisionEnabled(bIsCollisionEnabled)
+      MaterialNameList(InMaterialNameList),
+      bIsStationary(bInIsStationary),
+      bIsPhysicsEnabled(bInIsPhysicsEnabled),
+      bIsCollisionEnabled(bInIsCollisionEnabled)
 {
+    bIsTickEnabled = false;
+    bIsSelfCollision = false;
 }
 
-void FRRActorSpawnInfo::operator()(const FString& InUniqueName,
-                                   const FTransform& InTransform,
-                                   const FString& InMeshUniqueName,
-                                   const TArray<FString>& InMeshMaterialNameList,
-                                   bool bIsStationary,
-                                   bool bIsPhysicsEnabled,
-                                   bool bIsCollisionEnabled)
+void FRRActorSpawnInfo::operator()(const FString& InEntityModelName,
+                                   const FString& InUniqueName,
+                                   const FTransform& InActorTransform,
+                                   const TArray<FTransform>& InMeshRelTransformList,
+                                   const TArray<FString>& InMeshUniqueNameList,
+                                   const TArray<FString>& InMaterialNameList,
+                                   bool bInIsStationary,
+                                   bool bInIsPhysicsEnabled,
+                                   bool bInIsCollisionEnabled)
 {
-    // Take in only if either one is NOT EMPTY!
-    if (!InMeshUniqueName.IsEmpty())
-    {
-        MeshUniqueNameList.Add(InMeshUniqueName);
-    }
+    EntityModelName = InEntityModelName;
+    UniqueName = InUniqueName;
+
+    MeshUniqueNameList = InMeshUniqueNameList;
+    MeshUniqueNameList.Remove(EMPTY_STR);
+    MeshRelTransformList = InMeshRelTransformList;
 
     // Even if this actor has mesh content, it could use the [StaticMesh]'s built-in materials.
-    // Thus, [MeshMaterialNameList], which is configured to be passed in at [ARRMeshActor] spawning,
+    // Thus, [MaterialNameList], which is configured to be passed in at [ARRMeshActor] spawning,
     // does NOT NECESSARILY always contain some material!
-    MeshMaterialNameList = InMeshMaterialNameList;
-    MeshMaterialNameList.Remove(EMPTY_STR);
+    MaterialNameList = InMaterialNameList;
+    MaterialNameList.Remove(EMPTY_STR);
 
-    UniqueName = InUniqueName;
-    Transform = InTransform;
-
-    IsStationary = bIsStationary;
-    IsPhysicsEnabled = bIsPhysicsEnabled;
-    IsCollisionEnabled = bIsCollisionEnabled;
+    bIsStationary = bInIsStationary;
+    bIsPhysicsEnabled = bInIsPhysicsEnabled;
+    bIsCollisionEnabled = bInIsCollisionEnabled;
+    bIsTickEnabled = false;
+    bIsSelfCollision = false;
 }
 
 // ===================================================================================================================================
@@ -151,15 +146,20 @@ URRActorCommon* URRActorCommon::GetActorCommon(int8 InSceneInstanceId, UClass* I
         actorCommon->SceneInstanceId = InSceneInstanceId;
         SActorCommonList.Add(InSceneInstanceId, actorCommon);
 
-        // This [OnWorldCleanup()] is meant to called once only!
+        // This [OnPostWorldCleanup()] is meant to called once only!
         // Not sure about whether there is an UE internal check on already bound function, but better not relying on it.
         if (URRActorCommon::DEFAULT_SCENE_INSTANCE_ID == InSceneInstanceId)
         {
-            FWorldDelegates::OnPostWorldCleanup.AddStatic(&URRCoreUtils::OnWorldCleanup);
+            FWorldDelegates::OnPostWorldCleanup.AddStatic(&URRActorCommon::OnPostWorldCleanup);
         }
     }
 
     return SActorCommonList[InSceneInstanceId];
+}
+
+void URRActorCommon::OnPostWorldCleanup(UWorld* World, bool bSessionEnded, bool bCleanupResources)
+{
+    URRActorCommon::SActorCommonList.Reset();
 }
 
 URRActorCommon::URRActorCommon()
@@ -192,9 +192,26 @@ void URRActorCommon::SetupEnvironment()
     // Here, we initialize artifacts that utilized play resources
     UWorld* currentWorld = GetWorld();
     checkf(currentWorld, TEXT("[URRActorCommon::SetupEnvironment] Failed fetching Game World"));
+
+    // Fetch Main background environment
+    if (!MainEnvironment)
+    {
+        MainEnvironment = URRUObjectUtils::FindEnvironmentActor(currentWorld);
+        // Not all maps has MainEnvironment setup
+    }
 }
 
 bool URRActorCommon::HasInitialized(bool bIsLogged) const
 {
     return true;
+}
+
+void URRActorCommon::MoveEnvironmentToSceneInstance(int8 InSceneInstanceId)
+{
+    if (IsValid(MainEnvironment))
+    {
+        const FVector currentEnvLocation = MainEnvironment->GetActorLocation();
+        const FVector sceneInstanceLocation = URRCoreUtils::GetSceneInstanceLocation(InSceneInstanceId);
+        MainEnvironment->SetActorLocation(FVector(sceneInstanceLocation.X, sceneInstanceLocation.Y, currentEnvLocation.Z));
+    }
 }