+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog for RapyutaSimulationPlugins repository
 
+## 0.0.3 ##
+* Add Assimp as 3rd dependency (release sources + lib)
+* Add URRProceduralMeshComponent & mesh util classes: `FRRMeshData, URRMeshUtils`
+
+## 0.0.2 ##
+* Add core sources (`URRActorCommon, URRObjectCommon, ARRBaseActor, ARRMeshActor, ARRSceneDirector, URRStaticMeshComponent, ARRGameState, URRGameInstance, ARRPlayerController, ~Utils`)
+* `ARRRobotVehicle`: Add `InitMoveComponent()`, allowing `RobotVehicleMoveComponent` to be custom-initialized in child classes
+* `URRGameSingleton`: Add skeletal mesh related resource types (`USkeletalMesh, USkeleton, UPhysicsAsset`) & their fetching apis
+
 ## 0.0.1 ##
-------------------------
 * Add CHANGELOG
 * Update maintainers